librbd: move object map async ops to standalone classes
In preparation for creating new async snapshot-related object map
operations, move the existing object map async ops to there own
namespace.
