+++ src/librbd/ObjectMap.h
@@ -7,7 +7,6 @@
 #include "include/rados/librados.hpp"
 #include "include/rbd/object_map_types.h"
 #include "common/bit_vector.hpp"
-#include "librbd/AsyncRequest.h"
 #include <boost/optional.hpp>
 
 class Context;
@@ -60,89 +59,12 @@ public:
   bool enabled() const;
 
 private:
-
-  class Request : public AsyncRequest {
-  public:
-    Request(ImageCtx &image_ctx, uint64_t snap_id, Context *on_finish)
-      : AsyncRequest(image_ctx, on_finish), m_snap_id(snap_id),
-        m_state(STATE_REQUEST)
-    {
-    }
-
-  protected:
-    const uint64_t m_snap_id;
-
-    virtual bool should_complete(int r);
-    virtual int filter_return_code(int r) const {
-      // never propagate an error back to the caller
-      return 0;
-    }
-    virtual void finish(ObjectMap *object_map) = 0;
-  private:
-    /**
-     * <start> ---> STATE_REQUEST ---> <finish>
-     *                   |                ^
-     *                   v                |
-     *            STATE_INVALIDATE -------/
-     */
-    enum State {
-      STATE_REQUEST,
-      STATE_INVALIDATE
-    };
-
-    State m_state;
-
-    bool invalidate();
-  };
-
-  class ResizeRequest : public Request {
-  public:
-    ResizeRequest(ImageCtx &image_ctx, uint64_t snap_id, uint64_t new_size,
-		  uint8_t default_object_state, Context *on_finish)
-      : Request(image_ctx, snap_id, on_finish), m_num_objs(0),
-        m_new_size(new_size), m_default_object_state(default_object_state)
-    {
-    }
-
-    virtual void send();
-  protected:
-    virtual void finish(ObjectMap *object_map);
-  private:
-    uint64_t m_num_objs;
-    uint64_t m_new_size;
-    uint8_t m_default_object_state;
-  };
-
-  class UpdateRequest : public Request {
-  public:
-    UpdateRequest(ImageCtx &image_ctx, uint64_t snap_id,
-                  uint64_t start_object_no, uint64_t end_object_no,
-                  uint8_t new_state,
-                  const boost::optional<uint8_t> &current_state,
-		  Context *on_finish)
-      : Request(image_ctx, snap_id, on_finish),
-        m_start_object_no(start_object_no), m_end_object_no(end_object_no),
-        m_new_state(new_state), m_current_state(current_state)
-    {
-    }
-
-    virtual void send();
-  protected:
-    virtual void finish(ObjectMap *object_map);
-  private:
-    uint64_t m_start_object_no;
-    uint64_t m_end_object_no;
-    uint8_t m_new_state;
-    boost::optional<uint8_t> m_current_state;
-  };
-
   ImageCtx &m_image_ctx;
   ceph::BitVector<2> m_object_map;
   uint64_t m_snap_id;
   bool m_enabled;
 
   void invalidate(uint64_t snap_id);
-  void resize(uint64_t num_objs, uint8_t default_state);
 };
 
 } // namespace librbd
